moved solution and work to primitives lib
