+++ energiminer/mineplant.cpp
@@ -9,7 +9,7 @@
 
 #include "libegihash-cl/OpenCLMiner.h"
 #include "energiminer/miner.h"
-#include "energiminer/work.h"
+#include "primitives/work.h"
 #include "energiminer/CpuMiner.h"
 #include "energiminer/TestMiner.h"
